#71 Allow autoscaling configuration for Azure/OpenStack
